+++ yash-semantics/src/expansion.rs
@@ -793,11 +793,11 @@ mod tests {
     #[test]
     fn from_error_for_message() {
         let code = Rc::new(Code {
-            value: "".to_string().into(),
+            value: "hello".to_string().into(),
             start_line_number: NonZeroU64::new(1).unwrap(),
             source: Source::Unknown,
         });
-        let location = Location { code, index: 0 };
+        let location = Location { code, range: 2..4 };
         let new_value = Variable {
             value: Value::Scalar("value".into()),
             last_assigned_location: Some(Location::dummy("assigned")),