+++ yash-syntax/src/parser/and_or.rs
@@ -128,6 +128,6 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "foo &&");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 6);
+        assert_eq!(e.location.range, 6..6);
     }
 }