+++ yash-syntax/src/parser/function.rs
@@ -154,7 +154,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "( ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 2);
+        assert_eq!(e.location.range, 2..2);
     }
 
     #[test]
@@ -176,7 +176,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "( ) ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 4);
+        assert_eq!(e.location.range, 4..4);
     }
 
     #[test]
@@ -198,7 +198,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "() foo ; ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 3);
+        assert_eq!(e.location.range, 3..6);
     }
 
     #[test]
@@ -313,6 +313,6 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "()b");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 2);
+        assert_eq!(e.location.range, 2..3);
     }
 }