+++ yash-syntax/src/parser/grouping.rs
@@ -136,12 +136,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), " { oh no ");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 1);
+            assert_eq!(opening_location.range, 1..2);
         });
         assert_eq!(*e.location.code.value.borrow(), " { oh no ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 9);
+        assert_eq!(e.location.range, 9..9);
     }
 
     #[test]
@@ -155,7 +155,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "{ }");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 2);
+        assert_eq!(e.location.range, 2..3);
     }
 
     #[test]
@@ -225,12 +225,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), " ( oh no");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 1);
+            assert_eq!(opening_location.range, 1..2);
         });
         assert_eq!(*e.location.code.value.borrow(), " ( oh no");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 8);
+        assert_eq!(e.location.range, 8..8);
     }
 
     #[test]
@@ -244,6 +244,6 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "( )");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 2);
+        assert_eq!(e.location.range, 2..3);
     }
 }