+++ yash-syntax/src/parser/lex/backquote.rs
@@ -110,7 +110,7 @@ mod tests {
         let result = block_on(lexer.backquote()).unwrap().unwrap();
         assert_matches!(result, TextUnit::Backquote { content, location } => {
             assert_eq!(content, []);
-            assert_eq!(location.index, 0);
+            assert_eq!(location.range, 0..2);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
@@ -134,7 +134,7 @@ mod tests {
                     BackquoteUnit::Literal('o')
                 ]
             );
-            assert_eq!(location.index, 0);
+            assert_eq!(location.range, 0..6);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
@@ -163,7 +163,7 @@ mod tests {
                     BackquoteUnit::Literal('\'')
                 ]
             );
-            assert_eq!(location.index, 0);
+            assert_eq!(location.range, 0..15);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
@@ -193,7 +193,7 @@ mod tests {
                     BackquoteUnit::Literal('\'')
                 ]
             );
-            assert_eq!(location.index, 0);
+            assert_eq!(location.range, 0..15);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
@@ -212,7 +212,7 @@ mod tests {
                 content,
                 [BackquoteUnit::Literal('a'), BackquoteUnit::Literal('b')]
             );
-            assert_eq!(location.index, 0);
+            assert_eq!(location.range, 0..12);
         });
 
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
@@ -231,12 +231,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "`");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 0);
+            assert_eq!(opening_location.range, 0..1);
         });
         assert_eq!(*e.location.code.value.borrow(), "`");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 1);
+        assert_eq!(e.location.range, 1..1);
     }
 
     #[test]
@@ -251,11 +251,11 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "`foo");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 0);
+            assert_eq!(opening_location.range, 0..1);
         });
         assert_eq!(*e.location.code.value.borrow(), "`foo");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 4);
+        assert_eq!(e.location.range, 4..4);
     }
 }