+++ yash-syntax/src/parser/simple_command.rs
@@ -235,12 +235,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "(a b");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 0);
+            assert_eq!(opening_location.range, 0..1);
         });
         assert_eq!(*e.location.code.value.borrow(), "(a b");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 4);
+        assert_eq!(e.location.range, 4..4);
     }
 
     #[test]
@@ -254,12 +254,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "(a;b)");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.index, 0);
+            assert_eq!(opening_location.range, 0..1);
         });
         assert_eq!(*e.location.code.value.borrow(), "(a;b)");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.index, 2);
+        assert_eq!(e.location.range, 2..3);
     }
 
     #[test]
@@ -297,7 +297,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "my=assignment");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.index, 0);
+        assert_eq!(sc.assigns[0].location.range, 0..13);
     }
 
     #[test]
@@ -315,19 +315,19 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.index, 0);
+        assert_eq!(sc.assigns[0].location.range, 0..2);
         assert_eq!(sc.assigns[1].name, "b");
         assert_eq!(sc.assigns[1].value.to_string(), "!");
         assert_eq!(*sc.assigns[1].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[1].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[1].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[1].location.index, 3);
+        assert_eq!(sc.assigns[1].location.range, 3..6);
         assert_eq!(sc.assigns[2].name, "c");
         assert_eq!(sc.assigns[2].value.to_string(), "X");
         assert_eq!(*sc.assigns[2].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[2].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[2].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[2].location.index, 7);
+        assert_eq!(sc.assigns[2].location.range, 7..10);
     }
 
     #[test]
@@ -514,7 +514,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a= ()");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.index, 0);
+        assert_eq!(sc.assigns[0].location.range, 0..2);
 
         let next = block_on(parser.peek_token()).unwrap();
         assert_eq!(next.id, Operator(OpenParen));
@@ -535,7 +535,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a=b()");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.index, 0);
+        assert_eq!(sc.assigns[0].location.range, 0..3);
 
         let next = block_on(parser.peek_token()).unwrap();
         assert_eq!(next.id, Operator(OpenParen));